requirements-sim.txt
httpx>=0.27
pyyaml>=6.0
python-dotenv>=1.0
matplotlib>=3.8
